Move get geojson method
